Computed steric sea level against depth
In this commit, the code is thoroughly restructured. The src directory now
contains scripts for reading and writing data, plotting output, computing steric
sea level (SSL) at a given time in the global ocean, the contribution to SSL
across depth sections, and the contribution to SSL against depth. The tests
directory now contains scripts for testing the functions in the scripts of the
src directory. For example, to plot the contribution to SSL against depth during
a historical year at a specific location, all we need to do is to execute the
appropriate script in the tests directory after providing it with the mesh,
input file, location (e.g. latitude and longitude) and time informations. The
test function in this script will call the relevant function(s) within the
script(s) of the src directory and serve our purpose. Some of the existing
scripts are renamed (and their descriptions modified), and the following scripts
are added:
(a) src/StericSeaLevelAgainstDepth.py for computing the contribution to steric
    sea level as a function of depth;
(b) tests/Test_StericSeaLevelAgainstDepth.py for testing the various functions
    of ../src/StericSeaLevelAgainstDepth.py.